% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPolar.R
\name{uPolar}
\alias{uPolar}
\title{uPolar Visualization tool for time series of microfluidic microscopic images}
\format{
uPolar(time_col,dist_col,area_col,adjust,refLine,track,RLS)
}
\usage{
uPolar(time_col, dist_col, area_col, adjust, refLine, track, RLS)
}
\arguments{
\item{time_col}{time array from dataset}

\item{dist_col}{distance array from dataset to calculate radius}

\item{area_col}{area array, if area is unavailable set to NA}

\item{adjust}{Cell area adjustment for over size cells (better visualization), use  'NA' if area dataset is not available}

\item{refLine}{reference line to  for cell alignment monitoring  use  'NA'  for default}

\item{track}{color code tracking for cells movement and division ( FALSE = no tracking , TRUE = tracking)}

\item{RLS}{RLS  array from dataset for each time-point (division happened = 1, no division = 0), use  'NA'  if RLS data is not availablle}
}
\value{
Microfludics Time Series ploar plot (RLS optional)
}
\description{
AN easy-to-use visualizing tool for Microfluidic based microscopy images to monitor cellular events in biomedical time-series data.
}
\details{
Packages :     devtools::install.packages(c("dplyr,plotly,plyr"))
}
\examples{
#  df= read.csv("../data/BC8_tp10.csv")
#  rls.tp10 = read.csv("../data/BC8_RLS_tp10.csv")
#  uPolar(df$time_num,df$dist,df$area,0.05,34,FALSE,rls.tp10$com)

}
