% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPolar.R
\name{uPolar}
\alias{uPolar}
\title{uPolar An Interactive 2D Visualization Tool for Microscopic Time-series Images}
\format{
uPolar(df,refPoint=NULL,title=NULL,refLine=NULL,Area=FALSE,aAdjust=NULL,zeroAdjust= NULL,track = FALSE,RLS= FALSE)
}
\usage{
uPolar(
  df,
  refPoint = NULL,
  title = NULL,
  refLine = NULL,
  Area = FALSE,
  aAdjust = NULL,
  zeroAdjust = NULL,
  track = FALSE,
  RLS = FALSE
)
}
\arguments{
\item{df}{dataset (CSV format), Features be named : Time, X, Y, Area (optional), RLS (optional)}

\item{refPoint}{Given numerical image reference point to calculate distance, format: c(xr,yr), default (refPoint= NULL) c(0,0)}

\item{title}{Plot title, default (title = NULL)}

\item{refLine}{numerical value for reference line (red) on the plot, default (refLine = NULL)}

\item{Area}{Area = FALSE (no area data available) , Area =TRUE (area data available)}

\item{aAdjust}{numerical value for adjust cell size by multiplication, default (aAdjust = NULL)}

\item{zeroAdjust}{numerical value Adjust minimum cell distance from plot center point, default (zeroPoint = NULL)}

\item{track}{color tag tracking for cells movement and division, trak = FALSE (no tag tracking), track = TRUE (tag tracking)}

\item{RLS}{RLS = FALSE (no RLS data available), RLS = TRUE (RLS data available)}
}
\value{
Microfludics Time Series uploar plot
}
\description{
AN easy-to-use visualizing tool for Microfluidic based microscopy images to monitor cellular events in biomedical time-series data.
}
\details{
Packages :     devtools::install.packages(c("dplyr,plotly,plyr"))
}
\examples{
#  df= read.csv("../data/BC8_tp10_TpTmXYAreaRLS.csv")
#  uPolar(df,refPoint=NULL,title=NULL,refLine=NULL,Area=FALSE,aAdjust=NULL,
#        zeroAdjust= NULL,track = FALSE,RLS= FALSE)

}
