% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPolar.R
\name{uPolar}
\alias{uPolar}
\title{μPolar Visualization tool for time series of microfluidic microscopic images}
\usage{
uPolar(df, time_col, obj_col, area_col, dist_col, offset, c.Adjust, num_images)
}
\description{
μPolar Visualization tool for time series of microfluidic microscopic images
}
