% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPolar.R
\name{uPolar}
\alias{uPolar}
\title{uPolar Visualization tool for time series of microfluidic microscopic images}
\format{uPolar(time_col,obj_col,dist_col,num_number,area_col,offset,c.Adjust,RLS)}
\usage{
uPolar(
  df,
  time_col,
  obj_col,
  dist_col,
  num_images,
  area_col,
  offset,
  c.Adjust,
  RLS
)
}
\arguments{
\item{df}{Dataset, features order ( Time for each image , Total cell number, cell area (optional), cell distance )}

\item{time_col}{Time number for each image}

\item{obj_col}{Total cells number at each image}

\item{dist_col}{Cell distance from reference point to calculate radius}

\item{num_images}{Total number of images}

\item{area_col}{Individual cell area, if area is unavailable set to NA}

\item{offset}{Raduis between maximum cell raduis and plolar plot outlayer line (better visualization),  use  'NA'  for default}

\item{c.Adjust}{Cell area adjustment for over size cells (better visualization), use  'NA' if area dataset is not available}

\item{RLS}{RLS dataset for each time-point (division happened = 1, no division = 0), use  'NA'  if RLS data is not availablle}
}
\value{
Microfludics Time Series ploar plot (RLS optional)
}
\description{
AN easy-to-use visualizing tool for Microfluidic based microscopy images to monitor cellular events in biomedical time-series data.
}
\details{
Packages :     devtools::install.packages(c("dplyr,plotly,plyr"))

ibrary(plotly)

library(dplyr)

library(plyr)
}
